pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "google-genai>=1.50.1",
    "litellm>=1.79.3",
    "psutil>=7.1.3",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]
